+++ content/graph/Misc/DirectedMstKACTL.h
@@ -8,11 +8,11 @@
 
 #include "DSU.h"
 
-struct Edge { int a, b; ll w; };
+struct Edge { int a, b; int w; };
 struct Node { /// lazy skew heap node
 	Edge key;
 	Node *l, *r;
-	ll delta;
+	int delta;
 	void prop() {
 		key.w += delta;
 		if (l) l->delta += delta;
@@ -30,11 +30,11 @@ Node *merge(Node *a, Node *b) {
 }
 void pop(Node*& a) { a->prop(); a = merge(a->l, a->r); }
 
-ll dmst(int n, int r, vector<Edge>& g) {
+int dmst(int n, int r, vector<Edge>& g) {
 	DSU dsu; dsu.init(n);
 	vector<Node*> heap(n);
 	trav(e, g) heap[e.b] = merge(heap[e.b], new Node{e});
-	ll res = 0;
+	int res = 0;
 	vi seen(n, -1), path(n); seen[r] = r;
 	F0R(s,n) {
 		int u = s, qi = 0, w;
